backend/requirements.txt
Django==5.2.10
djangorestframework
django-cors-headers
djangorestframework-simplejwt
django-filter
drf-spectacular
Pillow
